requirements.txt
Pillow>=10.0.0
pytest>=8.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.9
